Refactor +ou:dejs:format
Old version of 'ou' was hacky and not in keeping with the rest of the
arms in dejs:format. New version is like 'ot' but with defaults. Also
added a version to dejs-soft:format. Removed 'uf' and 'un' as well,
since they were only used as helpers to the old 'ou'.